worker/module.ts: renames the file
